axd-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pandas
requests
sqlalchemy
pymysql
cryptography
